+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.7.11]
 
 - added export setting "objectscript.export.addCategory" if enabled uses previous behavior, adds category folder to export folder, disabled by default
+- added Server actions menu, by clicking on server info from status bar. Open Management portal, Class Reference and toggle connection.
 
 ## [0.7.10]